requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
openai==1.3.5
anthropic==0.7.8
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
pandas==2.1.3
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1